Baseline with multiple flexible choices

+++ samples/csharp/getting-started/DeepLearning_TensorFlow_TransferLearning/ImageClassification.Train/DataModels/ImagePrediction.cs
@@ -11,6 +11,6 @@ namespace ImageClassification.DataModels
         public float[] Score;
 
         [ColumnName("PredictedLabel")]
-        public Int64 PredictedLabel;
+        public UInt32 PredictedLabel;
     }
 }
